Move `BatchProxy` to `stateful`
