+++ packages/engine/lib/stateful/Cargo.toml
@@ -6,5 +6,6 @@ edition = "2021"
 [dependencies]
 memory = { path = "../memory" }
 arrow = { version = "10.0.0", default-features = false }
+parking_lot = "0.11.1"
 thiserror = "1.0.21"
 tracing = "0.1.29"
