+++ packages/engine/lib/stateful/src/lib.rs
@@ -1,6 +1,7 @@
 // TODO: DOC: Describing this crate containing stateful things like agents, context, etc. and the
 //            `field` interface
 pub mod field;
+pub mod proxy;
 
 mod error;
