+++ packages/engine/src/datastore/table/pool/mod.rs
@@ -5,12 +5,12 @@ pub mod proxy;
 use std::sync::Arc;
 
 use parking_lot::RwLock;
+use stateful::{
+    proxy::{BatchReadProxy, BatchWriteProxy},
+    Result,
+};
 
 use self::proxy::{PoolReadProxy, PoolWriteProxy};
-use crate::datastore::{
-    error::Result,
-    table::proxy::{BatchReadProxy, BatchWriteProxy},
-};
 
 /// Internal trait to implement `BatchPool` without leaking `Arc<RwLock<B>>`
 trait Pool<B> {
